Module Storage.Roll

module Owner : Tezos_raw_protocol_008_PtEdo2Zk.Storage_sigs.Indexed_data_snapshotable_storage with type key = Tezos_raw_protocol_008_PtEdo2Zk.Roll_repr.t and type snapshot = Tezos_raw_protocol_008_PtEdo2Zk.Cycle_repr.t * int and type value = Tezos_protocol_environment_008_PtEdo2Zk__Environment.Signature.Public_key.t and type t := Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t

val clear : Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t -> Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t Tezos_protocol_environment_008_PtEdo2Zk__Environment.Lwt.t

module Next : Tezos_raw_protocol_008_PtEdo2Zk.Storage_sigs.Single_data_storage with type value = Tezos_raw_protocol_008_PtEdo2Zk.Roll_repr.t and type t := Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t

The next roll to be allocated.

module Limbo : Tezos_raw_protocol_008_PtEdo2Zk.Storage_sigs.Single_data_storage with type value = Tezos_raw_protocol_008_PtEdo2Zk.Roll_repr.t and type t := Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t

Head of the linked list of rolls in limbo

module Delegate_roll_list : Tezos_raw_protocol_008_PtEdo2Zk.Storage_sigs.Indexed_data_storage with type key = Tezos_protocol_environment_008_PtEdo2Zk__Environment.Signature.Public_key_hash.t and type value = Tezos_raw_protocol_008_PtEdo2Zk.Roll_repr.t and type t := Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t

Rolls associated to contracts, a linked list per contract

module Successor : Tezos_raw_protocol_008_PtEdo2Zk.Storage_sigs.Indexed_data_storage with type key = Tezos_raw_protocol_008_PtEdo2Zk.Roll_repr.t and type value = Tezos_raw_protocol_008_PtEdo2Zk.Roll_repr.t and type t := Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t

Use this to iter on a linked list of rolls

module Delegate_change : Tezos_raw_protocol_008_PtEdo2Zk.Storage_sigs.Indexed_data_storage with type key = Tezos_protocol_environment_008_PtEdo2Zk__Environment.Signature.Public_key_hash.t and type value = Tezos_raw_protocol_008_PtEdo2Zk.Tez_repr.t and type t := Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t

The tez of a contract that are not assigned to rolls

module Snapshot_for_cycle : Tezos_raw_protocol_008_PtEdo2Zk.Storage_sigs.Indexed_data_storage with type key = Tezos_raw_protocol_008_PtEdo2Zk.Cycle_repr.t and type value = int and type t := Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t

Index of the randomly selected roll snapshot of a given cycle.

module Last_for_snapshot : Tezos_raw_protocol_008_PtEdo2Zk.Storage_sigs.Indexed_data_storage with type key = int and type value = Tezos_raw_protocol_008_PtEdo2Zk.Roll_repr.t and type t = Tezos_raw_protocol_008_PtEdo2Zk.Raw_context.t * Tezos_raw_protocol_008_PtEdo2Zk.Cycle_repr.t

Last roll in the snapshoted roll allocation of a given cycle.