Module Tezos_raw_protocol_008_PtEdo2Zk.Fees_storage

type Tezos_protocol_environment_008_PtEdo2Zk.Error_monad.error +=

| Cannot_pay_storage_fee

type Tezos_protocol_environment_008_PtEdo2Zk.Error_monad.error +=

| Operation_quota_exceeded

type Tezos_protocol_environment_008_PtEdo2Zk.Error_monad.error +=

| Storage_limit_too_high

val origination_burn : Raw_context.t -> (Raw_context.t * Tez_repr.t) Tezos_protocol_environment_008_PtEdo2Zk.Error_monad.tzresult

Does not burn, only adds the burn to storage space to be paid

val record_paid_storage_space : Raw_context.t -> Contract_repr.t -> (Raw_context.t * Tezos_protocol_environment_008_PtEdo2Zk.Z.t * Tezos_protocol_environment_008_PtEdo2Zk.Z.t * Tez_repr.t) Tezos_protocol_environment_008_PtEdo2Zk.Error_monad.tzresult Tezos_protocol_environment_008_PtEdo2Zk.Lwt.t

The returned Tez quantity is for logging purpose only

val check_storage_limit : Raw_context.t -> storage_limit:Tezos_protocol_environment_008_PtEdo2Zk.Z.t -> unit Tezos_protocol_environment_008_PtEdo2Zk.Error_monad.tzresult

val start_counting_storage_fees : Raw_context.t -> Raw_context.t

val burn_storage_fees : Raw_context.t -> storage_limit:Tezos_protocol_environment_008_PtEdo2Zk.Z.t -> payer:Contract_repr.t -> Raw_context.t Tezos_protocol_environment_008_PtEdo2Zk.Error_monad.tzresult Tezos_protocol_environment_008_PtEdo2Zk.Lwt.t