Module Tezos_raw_protocol_008_PtEdo2Zk.Voting_period_repr

type kind =

| Proposal
protocols can be proposed

| Testing_vote
a proposal can be voted

| Testing
winning proposal is forked on a testnet

| Promotion_vote
activation can be voted

| Adoption
a delay before activation

The voting period kinds are ordered as follows: Proposal -> Testing_vote -> Testing -> Promotion_vote -> Adoption. This order is the one used be the function succ below.

val kind_encoding : kind Tezos_protocol_environment_008_PtEdo2Zk__Environment.Data_encoding.t

type voting_period = {

index : Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t;

kind : kind;

start_position : Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t;

}

A voting period can be of 5 kinds and is uniquely identified by a counter since the root.

type t = voting_period

type info = {

voting_period : t;

position : Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t;

remaining : Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t;

}

val root : start_position:Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t -> t

include Tezos_protocol_environment_008_PtEdo2Zk__Environment.Compare.S with type t := voting_period

type t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val encoding : t Tezos_protocol_environment_008_PtEdo2Zk__Environment.Data_encoding.t

val info_encoding : info Tezos_protocol_environment_008_PtEdo2Zk__Environment.Data_encoding.t

val pp : Tezos_protocol_environment_008_PtEdo2Zk__Environment.Format.formatter -> t -> unit

val pp_info : Tezos_protocol_environment_008_PtEdo2Zk__Environment.Format.formatter -> info -> unit

val pp_kind : Tezos_protocol_environment_008_PtEdo2Zk__Environment.Format.formatter -> kind -> unit

val reset : t -> start_position:Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t -> t

reset period ~start_position increment the index by one and set the kind to Proposal which is the period kind that start the voting process. start_position is the level at wich this voting_period started.

val succ : t -> start_position:Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t -> t

succ period ~start_position increment the index by one and set the kind to its successor. start_position is the level at which this voting_period started.

val position_since : Level_repr.t -> t -> Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t

val remaining_blocks : Level_repr.t -> t -> blocks_per_voting_period:Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t -> Tezos_protocol_environment_008_PtEdo2Zk__Environment.Int32.t