Module Tezos_raw_protocol_008_PtEdo2Zk__Lazy_storage_kind

Lazy_storage offers a unified interface for specific Michelson datatype that behave somewhat lazily, because they are intended to be quite big. Instead of serializing/deserializing the whole value to/from the storage, only an identifier is used. The identifier acts like a pointer. When using the value in a Michelson script, some part of it may be read from the storage, and a lightweight diff is computed. The diff is effectively applied to the storage at the end of the execution.

This module defines the different kinds of lazy storages and their basic properties. See also Lazy_storage_diff.

Lazy storage types are:

Big_map